Revert "ingresar nota 50%"
This reverts commit eb1773bcdf8bf9576bd081bfe9d721c5820154ad.
