+++ Recordate/src/app/auth/registro/registro.page.scss
@@ -18,8 +18,6 @@
     margin: 50px;
     color: white;
     font-size: 15px;
-  
-    
 }
 .chechbox_registro{
     margin-right: 20px;
