+++ Recordate/src/app/home-page/home-page.page.html
@@ -39,7 +39,7 @@
               <ion-card-header>
                 <ion-card-title class="titulo_cards">Titulo Nota</ion-card-title>
                 <ion-row>
-                <ion-col><img class="edit" src="assets/img/edit.png" alt="imagen de editar"routerLink="/note"></ion-col>  
+                <ion-col><img class="edit" src="assets/img/edit.png" alt="imagen de editar" [routerLink]="['/note']"></ion-col>  
                 <ion-col> <ion-checkbox class="edit" ></ion-checkbox> </ion-col>
               </ion-row>
               </ion-card-header>
