+++ Recordate/src/app/note/note.page.html
@@ -1,0 +1,64 @@
+
+<ion-header [translucent]="true">
+  <ion-toolbar>
+    <div class="container-text"><ion-title class="title-pag"><p class="title-pag">Recordatorios</p></ion-title></div>
+    <ion-button [routerLink]="['/home-page']" class="CancelarBoton">
+      Cancelar
+    </ion-button>
+    <ion-button [routerLink]="['/home-page']" class="AceptaBoton">
+      Guardar
+
+    </ion-button>
+  </ion-toolbar>
+</ion-header>
+<ion-content [fullscreen]="true">
+  <ion-header collapse="condense">
+    <ion-toolbar>
+      <ion-title class="container-text"size="large">Recordatorios</ion-title>
+    </ion-toolbar>
+  </ion-header>
+
+<br><br>
+<!--INICIO recordatorio-->
+  <ion-list>
+    <ion-list>
+      <ion-item>
+        <ion-input label="Titulo" placeholder="Titulo"></ion-input>
+      </ion-item> 
+      <ion-item>
+        <ion-input label="Descripcion" placeholder="Descripcion" ></ion-input>
+      </ion-item>
+  <br><br>
+   <!--categoria recordatorio-->
+
+
+   <ion-list>
+    <ion-item>
+      <ion-select label="Perfil" placeholder="Elige Perfil">
+        <ion-select-option value="estudios">Estudios</ion-select-option>
+        <ion-select-option value="compras">Compras</ion-select-option>
+        <ion-select-option value="personal">Personal</ion-select-option>
+      </ion-select>
+    </ion-item>
+
+
+    <!--categoria recordatorio-->
+    </ion-list>
+
+
+<ion-item>Elige fecha y hora</ion-item>
+
+  </ion-list>
+<!--FIN recordatorio-->
+
+  <!--Fecha y hora-->
+
+
+  <div class="containerFecha">
+    <ion-datetime  locale="en-GB" hourCycle="h12"></ion-datetime>
+  </div>
+
+  
+ <!--Fecha y hora-->
+
+
