+++ Recordate/src/app/note/note.page.scss
