+++ Recordate/src/app/note/note.page.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-note',
+  templateUrl: './note.page.html',
+  styleUrls: ['./note.page.scss'],
+})
+export class NotePage implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
