--- Recordate/src/app/pages/note/note.page.html
@@ -1,40 +1,0 @@
-<ion-header [translucent]="true">
-  <ion-toolbar>
-    <div class="container-text">
-      <ion-title class="title-pag">
-        <p class="title-pag">Recordatorios</p>
-      </ion-title>
-    </div>
-    <ion-button [routerLink]="['/home-page']" class="CancelarBoton">
-      Cancelar
-    </ion-button>
-    <ion-button (click)="submit()" class="AceptaBoton">
-      Guardar
-    </ion-button>
-  </ion-toolbar>
-</ion-header>
-
-<ion-content [fullscreen]="true">
-  <ion-header collapse="condense">
-    <ion-toolbar>
-      <ion-title class="container-text" size="large">Recordatorios</ion-title>
-    </ion-toolbar>
-  </ion-header>
-
-  <app-custom-input [control]="form.controls.nombreNota" type="text" label="Nombre" icon="person-outline"></app-custom-input>
-
-  <!-- Fecha y hora -->
-
-  <app-custom-input [control]="form.controls.descripcion" type="text" label="Descripción" icon="clipboard-outline"></app-custom-input>
-
-  <app-custom-input
-  [control]="form.controls.tipoNotaId"
-  type="select"
-  label="Tipo de nota"
-  icon="clipboard-outline"
-  [allowedOptions]="getAllowedOptions()"
-></app-custom-input>
-
-
-
-</ion-content>
