+++ Recordate/src/app/shared/custom-input/custom-input.component.ts
@@ -6,25 +6,30 @@ import { FormControl } from '@angular/forms';
   templateUrl: './custom-input.component.html',
   styleUrls: ['./custom-input.component.scss'],
 })
-export class CustomInputComponent implements OnInit {
+export class CustomInputComponent  implements OnInit {
 
   @Input() control: FormControl;
   @Input() label: string;
   @Input() icon: string;
   @Input() type: string;
-  @Input() allowedOptions: string[];
+  @Input() autocomplete: string;
+
   isPassword: boolean;
   hide: boolean = true;
 
   constructor() { }
 
   ngOnInit() {
-    if (this.type === 'password') {
-      this.isPassword = true;
-    }
+    if(this.type == 'password') this.isPassword = true;
   }
 
   showPassword() {
     this.hide = !this.hide;
+    if(this.hide) {
+      this.type = 'password'
+    }else{
+      this.type = 'text'
+    }
   }
+
 }
